Refactoring of unit tests
