+++ tests/unit/weather/test_weather_platform_co.py
